Moved inbox letters into their own component
