+++ src/shared/features/GrandparentViews/Inbox/Inbox.css
@@ -6,39 +6,6 @@
   height:calc(70vh - 50px);
 }
 
-.inboxLetter img {
-  display:block;
-  height:100%;
-  object-fit:contain;
-}
-
 .noInboxMargin.MuiGrid-item {
   margin-top:0px !important;
-}
-
-.inboxLetterInfo {
-  text-align:center;
-  height:52px !important;
-  background:transparent !important;
-}
-
-.inboxLetterInfo .MuiGridListTileBar-titleWrap .MuiGridListTileBar-title {
-  font-size:1.4rem;
-  line-height:32px;
-  color:black !important;
-}
-
-.inboxLetterInfo .MuiGridListTileBar-titleWrap .MuiGridListTileBar-subtitle {
-  font-size:1.2rem;
-  color:#666 !important;
-}
-
-@media screen and (max-width:500px) {
-  .inboxLetterInfo .MuiGridListTileBar-titleWrap .MuiGridListTileBar-title {
-    font-size:1.0rem;
-  }
-  
-  .inboxLetterInfo .MuiGridListTileBar-titleWrap .MuiGridListTileBar-subtitle {
-    font-size:0.8rem;
-  }
 }
